Fix render engine start signalling.
Instead of m_started we need to be setting m_available. If
we control m_started through this mechanism we'd have the
RenderedDisplayMode call repeatedly into StartRender() because
when it checks for IsRendererStarted in a pretty tight loop it'd
potentially see m_started still being false.

This very double/tripple call resulted in view updates not arriving
in the right place, and in potential crashes.

#RH-31966
#RH-31926