Use new QueueSingleView() to ensure Light conversion goes correct.
Before this change I was under the assumption that m_current_view_info
(ViewInfo) would always be valid, but the underlying C++ CQueue
implementation clears that out, so invalid after the Flush with a view
change ends.

Toggling skylighting off enables default lighting, but since I have no
longer a valid ViewInfo that would lead to bad location and direction
information after conversion of the passed in default Light.

This change builds on a set of changes needed for #RH-32345 and #RH-32356
that ensure I have access to a valid ViewInfo on which I can base the
default light conversion.

#RH-32345 Needs Testing
#RH-32356 Needs Testing