Skip linear to sRGB conversion for render results.
I have added an option to turn off the linear to sRGB
conversion that happened when rendering as progressive
foreground session.

Cycles has a tonemapping step, that was originally added
to ensure render results matched color space of Blender
UI, but that is not necessarily true when integrating
Cycles in other software, like RhinoCycles.

Thus the need to make it optional.