RhinoCycles: Aspect ratio of render result in RenderWindow wrong.
With a very wide, not so heigh viewport but rendering in a square
render window it was very apparent the aspect ratio was wrong.

The viewport size was used to calculate the camera settings, but
the render window size was used to render as, so render results
would show up squashed in above example.

Make sure we use the render setting defined sizes when doing modal
and preview renders instead.

#RH-32533 Needs Testing