Turn RhinoCycles into a utility plug-in.
This is in preparation for v6. The Rendered with Cycles
viewport mode will be available, but as Raytraced instead.
There will not be a RhinoCycles renderer entry through
this plug-in available. Instead a new plug-in will be
created in the near future that will essentially enable
the modal rendering part of this implementation.