Don't update display while waiting for a ChangeQueue flush to complete.
Allowing the display handler to go through it all we may run into memory
being deleted while it is still in use.

It could happen that the changequeue signalled us for changes, but we
start handling the display update just before we actually upload the
data to CCSycles. In our API layer we have buffers to contain pixel data.
On view size changes we through that buffer away and recreate one.

To prevent us from using it and throwing it away at the same time make
sure we don't use those buffers while we're uploading data.