RhinoCycles: create Cycles for Rhino -plugin.
Re-organise code so that the core code is shared between two
plug-ins: RhinoCycles.rhp and CyclesForRhino.rhp.

The core code is now available in the RhinoCyclesCore project
and compiles to RhinoCyclesCore.dll.

RhinoCycles.rhp stays a utility plug-in, providing the viewport mode
interactive rendering. This plug-in will always load at start-up, and
is thus responsible for initialising Cycles through CSycles.

CyclesForRhino.rhp is the new render plug-in. It provides the (semi-)
modal rendering functionality. Loading the plug-in adds a new render
engine entry called 'Cycles for Rhino', and will render to the
(semi-) modal render window, also known to users as the stand-alone
renderer.