Move to ViewInfo over RhinoView for ChangeQueue.
Explicit split between ViewInfo and ViewportInfo is necessary for supporting
ViewCaptureTo* that clones the display pipeline.

This way we can give the ChangeQueue the proper N_3dmView and N_Viewport it
needs for functioning. specially since ViewCaptureTo* result in creating
modified copies of those that don't congrue with the ones in the Rhino
document.

This means that render engines implementing the DisplayMode will need to
realise that and take care they don't assume everything comes from the
Rhino doc.

 #RH-32206