Simplify lock usage and event triggering.
Remove locks that aren't needed now that we're
handling most of the render window updating in
the render engine implementation instead of
the view mode implementation.

Also use the ?. operator in places where it is
suitable. This makes the code much more readable
by cutting down on unnecessary, explicit null-checks.