Use Rhino.Geometry for generating ground plane mesh.
Instead of painstakingly generating a ground plane
mesh along with UV coordinates use the geometry tools
available in RhinoCommon.

This allows us to have more freedom in placement of the
plane and controlling the size of it whilst keeping
UV mapping correct in case a proper material has been
assigned to the ground plane.

#RH-33545