Add a mechanism to specify a throttle for Raytraced viewport.
On some machines it can happen that the Raytraced viewport
slows down the user machine quite a bit. This is especially
so with resource intensive scenes where only one GPU is used,
or the GPU that is the primary device for the Windows desktop.

To alleviate problems arising from the full GPU usage the throttle
is introduced. It'll allow the OS to use the GPU too (: