Allow control of render result blending into draw buffer.
There is a fade in control that can be used to ease the raytraced
render result into the viewport. With higher values for
FadeInMs the ease will be slower. By default this is set to
10 ms.

UseFastDraw() is implemented to return true, meaning the underlying
pipeline doesn't wait for IsFramebufferAvailable anymore. As a
result the OpenGL rendered view should be available right away.