Add UseDrawOpenGl setting.
With this setting one can change the setting
for the registered display mode on next Rhino
start.

The default setting is true, meaning the OpenGL
method of drawing is used. Setting it to false
will put the viewport update to using the
RenderWindow interface. This interface suits
session over RDP (which is done automatically),
but also older equipment that doesn't have recent
enough OpenGL driver available (minimum for OpenGL
drawing is 3.3).