Fix incorrect environment setup for skylight.
The old logic meant that the bg environment
was being used for skylight effect. This is not
a problem when the 360 environment is also
used as the skylighting environment. The problem
is evident when the environment is set to a solid
color (i.e. white), but the skylight is set to
an environment with a lowered intensity.

The fix is to properly use light path combinations
to pick the right color. The correct tests are:

* choose between bg and refl environment color:
  use refl environment when IsGlossyRay *and*
  IsReflectionRay
* choose between bg and skylight environment:
  use bg environment when IsCameraRay