Blanket scene reset results in unnecessary data upload.
The scene reset was a remnant of old times where updates
were still pretty much hard to get right due to difficulties
in stopping Cycles for updates. The blanket reset on the
scene somehow gave some false sense of stability.

The scene reset is not needed, and can thus be removed.