Reorganize SignalRedraw() usage behind lock.
Instead of having different threads post a redraw
signal we now don't post new redraws when we still
have a redraw signal in progress - i.e. we have signalled
to redraw, but the redraw hasn't happened yet.

This seems to help alleviate the boging down problem
Raytracing has on the color picker.