Tweak materials somewhat.
metal: now goes fully through principled metallic flow.
plastic/glass: tinting (sheen and specular) is done such
that with the inverse of the reflectivity is used. That will
give a less white highlight when lower reflection is used. Note
that tinting is done towards the specified base color.
clearcoat is also used for plastic/glass, but for metallic
workflow it is completely turned off.