Attempt to protect against NullReferenceException.
Apparently it can happen that the ViewportRenderEngine instance
is disposed of while there is still code executing on it.

Springle checks around that see if 'this' is still valid and
hopefully allow for a more-or-less clean exit.

Additionally sprinkle checks for CancelRender and bail out as early
as possible. This should make cancelling a switch to Raytraced
much more responsive instead of having to wait until data has been
handled for a certain step during which the render session was
cancelled.