Add mechanism to detect known bad OpenCL hardware.
Allow to skip OpenCL usage when such hardware is found. Also
allow to modify the 'skip list' without having to recompile
Rhino. This is useful when new bad hardware is found and needs
to be worked around before a next SR is released.

The skip list is a file called `skiplist.xml` in the
SettingsDirectory.

It looks like follows

```
<ArrayOfGpuDevice xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
  <GpuDevice>
    <DeviceName>Intel</DeviceName>
    <SubStrings>
      <SubString>
        <Value>530</Value>
      </SubString>
    </SubStrings>
  </GpuDevice>
</ArrayOfGpuDevice>
```

ArrayOfGpuDevice contains one or more GpuDevice. Each
GpuDevice contains a SubStrings that contains zero or
more SubString

The above snippet will skip OpenCL usage when an the card
"Intel(r) HD Graphics 530" is found.