Special handling of amount when outside [0,1] range.
When a texture amount is outside of the [0,1] range
we need to separate the color channels and multiply
them with the amount to get the proper HDR colors.

One use-case example is subsurface scattering radius
where the amount can go well over 1.0. We need to be
able to get this radius info transfered properly.