Use multiplier when a texture is set in the Emission channel
Multiplier is actually factored into the emission color, so use
the component with the largest value as the emission strength.
This needs to be done explicitly for emission with texture
because otherwise it will stay just at 1.

No need to do that for just color emission, because color has
the multiplier factored in.